fix(v2): query frontends compatibility
